+++ src/app/landing-page/AnimatedShapes.tsx
