+++ src/app/landing-page/BrandsSection.tsx
