+++ src/app/landing-page/WhoWeAre.tsx
