+++ src/app/page.tsx
@@ -2,9 +2,9 @@ import React from 'react';
 import Navbar from '@/components/Navbar';
 import Footer from '@/components/Footer';
 import CallToAction from '@/components/CallToAction';
-import Hero from '@/components/Hero';
-import WhoWeAre from '@/components/WhoWeAre';
-import OurReach from '@/components/OurReach';
+import Hero from '@/app/landing-page/Hero';
+import WhoWeAre from '@/app/landing-page/WhoWeAre';
+import OurReach from '@/app/landing-page/OurReach';
 
 export default function LandingPage() {
   const reachStats = [
